requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
textblob
vaderSentiment
scikit-learn
joblib
